feat(M02-PO): author all ksp_PO_* read SPs and fix Dapper type mapping
- Add 15 new stored procedures (ksp_PO_Get*) covering all PO read
  endpoints: Parameters, UserLevel, PreAddChecks, OrderTypes, Suppliers,
  Carriers (PO_CAR), FormTypes (PO_FormType), GstTaxCodes, GSTRouting,
  Addresses, PRLines, LastPO, POHeader, Print
- Column names verified against live JAT schema (PO_ORDH, PO_ORDL,
  PO_ORDL_DETL, PO_CAR, PO_FormType, FA_SLMAS, pp_divmas, IG_TAX)
- Fix DeliverySlotRow record: LineNo int→decimal (PORDSNO is NUMERIC),
  SlotNo int→long (ROW_NUMBER() returns BIGINT); add casts in
  ReadDeliveryAsync to int for DTO mapping
- Update merged.sql with all SP changes (5017 lines)
- Switch VITE_API_URL to relative path for local dev proxy
- Clear hardcoded server URL from LoginPage display

+++ Development/Backend/Spinrise.Infrastructure/Areas/PurchaseOrder/PoEntry/PoEntryRepository.cs
@@ -300,8 +300,8 @@ public class PoEntryRepository : IPoEntryRepository
 
     private static async Task<List<DeliveryScheduleLineDto>> ReadDeliveryAsync(SqlMapper.GridReader multi)
     {
-        var slots     = (await multi.ReadAsync<DeliverySlotRow>()).ToList();
-        var slotsByLine = slots.GroupBy(s => s.LineNo).ToDictionary(g => g.Key, g => g.ToList());
+        var slots       = (await multi.ReadAsync<DeliverySlotRow>()).ToList();
+        var slotsByLine = slots.GroupBy(s => (int)s.LineNo).ToDictionary(g => g.Key, g => g.ToList());
 
         return slotsByLine.Select(kvp => new DeliveryScheduleLineDto
         {
@@ -313,7 +313,7 @@ public class PoEntryRepository : IPoEntryRepository
             PoQty    = kvp.Value.First().PoQty,
             Slots    = kvp.Value.Select(s => new DeliverySlotDto
             {
-                SlotNo  = s.SlotNo,
+                SlotNo  = (int)s.SlotNo,
                 ShDate  = s.ShDate,
                 Qty     = s.Qty,
                 Remarks = s.Remarks
@@ -322,13 +322,13 @@ public class PoEntryRepository : IPoEntryRepository
     }
 
     private sealed record DeliverySlotRow(
-        int     LineNo,
+        decimal LineNo,   // PORDSNO is NUMERIC in SQL — must be decimal
         string  ItemCode,
         string  ItemName,
         string  Uom,
         decimal PrNo,
         decimal PoQty,
-        int     SlotNo,
+        long    SlotNo,   // ROW_NUMBER() always returns BIGINT — must be long
         string? ShDate,
         decimal Qty,
         string  Remarks);
